# swagger/ecomm.yaml
openapi: 3.0.3
info:
  title: Swagger ECOMM - OpenAPI 3.0
  description: API usada para curso da Alura Level Up - Next Dev
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.11
servers:
  - url: https://localhost:3000
tags:
  - name: category
    description: Categorias de produtos 
  - name: product
    description: Produtos
  - name: user
    description: Usuarios
  - name: order
    description: Pedidos
paths:
  /api/categories:
    get:
      tags:
        - category
      summary: Get all categories
      description: Retorna todas as categorias com nome e status
      operationId: findCategorias
      responses: 
        '200':
          description: SUCESSO!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/categories"
        '404':
          description: Nenhuma categoria encontrada
          content:
            application/json:
              example: "Not Found"
  /api/categories/{id}:
    get:
      tags:
        - category
      summary: Get one category
      description: Retorna uma categoria especifica
      operationId: findCategoriaById
      parameters:
        - name: id
          in: path
          description: id da categoria
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        '200':
          description: SUCESSO!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/category"
        '404':
          description: Nenhuma categoria encontrada
          content:
            application/json:
              example: "Not Found"        
  /api/admin/categories:
    post:
      tags:
        - category
      summary: Create one category
      description: Cria uma categoria com nome e status
      operationId: createCategoria
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createCategory"
      responses:
        '201':
          description: Categoria criada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/category"
        '400':
          description: Erro ao criar categoria
          content:
            application/json:
              example: "Bad request"
  /api/admin/categories/update/{id}:
    put:
      tags:
        - category
      summary: Update a category
      description: Atualiza uma categoria
      operationId: updateCategoria
      parameters:
        - name: id
          in: path
          description: id da categoria
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createCategory"
      responses:
        '200':
          description: Categoria atualizada!!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/category"
        '404':
          description: Categoria não encontrada
          content:
            application/json:
              example: "Not found"
        '400':
          description: Requisição mal formatada
          content:
            application/json:
              example: "Bad Request"
  /api/admin/categories/remove/{id}:
    delete:
      tags:
        - category
      summary: Remove a category
      description: Remove uma categoria
      operationId: deleteCategoria
      parameters: 
        - name: id
          in: path
          description: id da categoria
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        '200': 
          description: Categoria removida
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/category"
        '404':
          description: Categoria não encontrada
          content:
            application/json:
              example: "Not found"
  /api/admin/categories/active/{id}:
    put:
      tags:
        - category
      summary: Active a category
      description: Ativa uma categoria já existente
      operationId: activeCategoria
      parameters:
        - name: id
          in: path
          description: id da categoria
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        '200': 
          description: Categoria ativada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/category"
        '404':
          description: Categoria não encontrada
          content:
            application/json:
              example: "Not found"
  /api/products:
    get:
      tags:
        - product
      summary: Get all products
      description: Retorna todos os produtos
      operationId: findProduts
      parameters:
        - name: id
          in: query
          description: id da categoria a ser filtrada
          required: false
          schema:
              oneOf:
                - type: string
                - type: integer
        - name: minEstoque
          in: query
          description: valor minimo de estoque de um produto
          required: false
          schema:
            type: integer
        - name: maxPreco
          in: query
          description: preço máximo de um produto
          required: false
          schema:
            type: integer
        - name: minPreco
          in: query
          description: preço mínimo de um produto
          required: false
          schema:
            type: integer
      responses: 
        '200':
          description: SUCESSO!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/products"
        '404':
          description: Nenhum produto encontrado
          content:
            application/json:
              example: "Not Found"
  /api/products/{id}:
    get:
      tags:
        - product
      summary: Get one product
      description: Retorna um produto especifico
      operationId: findProductById
      parameters:
        - name: id
          in: path
          description: id do produto
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        '200':
          description: SUCESSO!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/product"
        '404':
          description: Nenhum produto encontrado
          content:
            application/json:
              example: "Not Found"     
  /api/admin/products:
    post:
      tags:
        - product
      summary: Create one product
      description: Cria um produto
      operationId: createProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createProduct"
      responses:
        '201':
          description: Produto criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/product"
        '404':
          description: Erro ao criar produto
          content:
            application/json:
              example: "Not found"
  /api/admin/products/update/{id}:
    put:
      tags:
        - product
      summary: Update a product
      description: Atualiza um produto
      operationId: updateProduct
      parameters:
        - name: id
          in: path
          description: id do produto
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createProduct"
      responses:
        '200':
          description: Produto atualizado!!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/product"
        '404':
          description: Produto não encontrado
          content:
            application/json:
              example: "Not found"
        '400':
          description: Requisição mal formatada
          content:
            application/json:
              example: "Bad Request"
  /api/admin/products/remove/{id}:
    delete:
      tags:
        - product
      summary: Remove a product
      description: Remove um produto
      operationId: deleteProduct
      parameters: 
        - name: id
          in: path
          description: id do produto
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        '200': 
          description: Produto removido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/product"
        '404':
          description: Produto não encontrado
          content:
            application/json:
              example: "Not found"
        '400':
          description: Requisição mal formatada
          content:
            application/json:
              example: "Bad Request"
  /api/admin/users:
    get:
      tags:
        - user
      summary: Get all users
      description: Retorna todos os usuarios
      operationId: findUsers
      responses: 
        '200':
          description: SUCESSO!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/users"
        '404':
          description: Nenhum usuario encontrado
          content:
            application/json:
              example: "Not Found"
    post:
      tags:
        - user
      summary: Create one user
      description: Cria um usuario
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createUser"
      responses:
        '201':
          description: Usuario criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        '404':
          description: Erro ao criar usuario
          content:
            application/json:
              example: "Not found"
  /api/users/{id}:
    get:
      tags:
        - user
      summary: Get one user
      description: Retorna um usuario especifico
      operationId: findUserById
      parameters:
        - name: id
          in: path
          description: id do usuario
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        '200':
          description: SUCESSO!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        '404':
          description: Nenhum usuario encontrado
          content:
            application/json:
              example: "Not Found"     
  /api/admin/users/update/{id}:
    put:
      tags:
        - user
      summary: Update a user
      description: Atualiza um usuario
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: id do usuario
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createUser"
      responses:
        '200':
          description: Usuario atualizado!!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        '404':
          description: Usuario não encontrado
          content:
            application/json:
              example: "Not found"
        '400':
          description: Requisição mal formatada
          content:
            application/json:
              example: "Bad Request"
  /api/admin/users/remove/{id}:
    delete:
      tags:
        - user
      summary: Remove a user
      description: Remove um usuario
      operationId: deleteUser
      parameters: 
        - name: id
          in: path
          description: id do usuario
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        '200': 
          description: Usuario removido
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/user"
        '404':
          description: Usuario não encontrado
          content:
            application/json:
              example: "Not found"
        '400':
          description: Requisição mal formatada
          content:
            application/json:
              example: "Bad Request"
  /api/orders/{id}:
    get:
      tags:
        - order
      summary: Get one order
      description: Retorna um pedido especifico
      operationId: findOrderById
      parameters:
        - name: id
          in: path
          description: id do pedido
          required: true
          schema:
            oneOf:
              - type: string
              - type: integer
      responses:
        '200':
          description: SUCESSO!
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/order"
        '404':
          description: Nenhum pedido encontrado
          content:
            application/json:
              example: "Not Found"   
  /api/orders:
    post:
      tags:
        - order
      summary: Create one order
      description: Cria um pedido
      operationId: createOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/createOrder"
      responses:
        '201':
          description: Pedido criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/order"
        '404':
          description: Erro ao criar pedido
          content:
            application/json:
              example: "Not found"
components:
  schemas:
    category:
      type: object
      properties:
        _id:
          oneOf:
            - type: string
            - type: integer
        nome:
          type: string
        status:
          type: string
    categories:
      type: array
      items:
        $ref: "#/components/schemas/category"
    createCategory:
      type: object
      properties:
        nome: 
          type: string
        status: 
          type: string
    product:
      type: object
      properties:
        _id:
          oneOf:
            - type: string
            - type: integer
        nome:
          type: string
        descricao:
          type: string
        slug:
          type: string
        preco:
          type: number
        estoque:
          type: integer
        categoria:
          type: string
    products:
      type: array
      items:
        $ref: "#/components/schemas/product"
    createProduct:
      type: object
      properties:
        nome:
          type: string
        descricao:
          type: string
        slug:
          type: string
        preco:
          type: number
        estoque:
          type: integer
        categoria:
          type: string
    user:
      type: object
      properties:
        _id:
          oneOf:
            - type: string
            - type: integer
        username:
          type: string
        email:
          type: string
        senha:
          type: string
        cpf:
          type: string
        telefone:
          type: string
        dataCriacao:
          type: string
        endereco:
          type: object
          properties:
            bairro:
              type: string
            rua:
              type: string
            numero:
              oneOf:
                - type: string
                - type: integer
            complemento:
              type: string
            cep:
              type: string
            cidade:
              type: string
            uf:
              type: string
    users: 
      type: array
      items:
        $ref: "#/components/schemas/user"
    createUser:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        senha:
          type: string
        cpf:
          type: string
        telefone:
          type: string
        endereco:
          type: object
          properties:
            bairro:
              type: string
            rua:
              type: string
            numero:
              oneOf:
                - type: string
                - type: integer
            complemento:
              type: string
            cep:
              type: string
            cidade:
              type: string
            uf:
              type: string
    order:
      type: object
      properties:
        _id:
          oneOf:
            - type: string
            - type: integer
        dataPedido:
          type: string
        account:
          type: object
          properties:
            accountId:
              oneOf:
                - type: string
                - type: integer
            nomeCliente: 
              type: string
        enderecoEntrega:
          type: object
          properties:
            bairro:
              type: string
            rua:
              type: string
            numero:
              oneOf:
                - type: string
                - type: integer
            complemento:
              type: string
            cep:
              type: string
            cidade:
              type: string
            uf:
              type: string
        itens: 
          type: array
          items: 
            type: object
            properties:
              productId:
                oneOf: 
                  - type: string
                  - type: integer
              quantidade:
                type: integer
              desconto:
                type: number
              precoUnitario:
                type: number
    createOrder:
      type: object
      properties:
        account:
          type: object
          properties:
            accountId:
              oneOf:
                - type: string
                - type: integer
            nomeCliente: 
              type: string
        enderecoEntrega:
          type: object
          properties:
            bairro:
              type: string
            rua:
              type: string
            numero:
              oneOf:
                - type: string
                - type: integer
            complemento:
              type: string
            cep:
              type: string
            cidade:
              type: string
            uf:
              type: string
        itens: 
          type: array
          items: 
            type: object
            properties:
              productId:
                oneOf: 
                  - type: string
                  - type: integer
              quantidade:
                type: integer
              desconto:
                type: number
              precoUnitario:
                type: number